Only delete cache if cache folder exist

+++ autosubliminal/core/cache.py
@@ -45,7 +45,8 @@ class MutexFileLock(AbstractFileLock):
 
 def clear_mako_cache():
     mako_cache_dir = os.path.abspath(os.path.join(autosubliminal.CACHEDIR, 'mako'))
-    shutil.rmtree(mako_cache_dir, onerror=set_rw_and_remove)
+    if os.path.exists(mako_cache_dir):
+        shutil.rmtree(mako_cache_dir, onerror=set_rw_and_remove)
 
 
 def clear_imdbpie_cache():
@@ -53,7 +54,8 @@ def clear_imdbpie_cache():
     # Cleanup is required when switching between python versions
     # If not, 'ValueError: unsupported pickle protocol' is thrown
     imdb_cache_file = os.path.abspath(os.path.join(tempfile.gettempdir(), 'cache.db'))
-    os.remove(imdb_cache_file)
+    if os.path.exists(imdb_cache_file):
+        os.remove(imdb_cache_file)
 
 
 def cache_artwork(indexer_name, indexer_id, artwork_type, artwork_url, thumbnail=False):
